+++ app/core/ai.py
@@ -11,10 +11,8 @@ The module uses a global cache to store initialized agents for better performanc
 """
 
 import logging
-import os
 import time
 
-from anyio.lowlevel import checkpoint
 from cdp_langchain.agent_toolkits import CdpToolkit
 from cdp_langchain.utils import CdpAgentkitWrapper
 from fastapi import HTTPException
@@ -27,11 +25,11 @@ from sqlalchemy.exc import SQLAlchemyError
 from sqlalchemy.orm.exc import NoResultFound
 from twitter_langchain import TwitterApiWrapper, TwitterToolkit
 
-from app.config import config
-from app.db import Agent, get_coon, get_db
-from skill.common import get_common_skill
-from skill.crestal import get_crestal_skill
-from skill_set import get_skill_set
+from app.config.config import config
+from app.models.db import Agent, get_coon, get_db
+from skills.common import get_common_skill
+from skills.crestal import get_crestal_skill
+from skill_sets import get_skill_set
 
 logger = logging.getLogger(__name__)
