+++ app/entrypoints/autonomous.py
@@ -1,15 +1,14 @@
 import logging
 import signal
 import sys
-import time
 from datetime import datetime, timedelta
 
 from apscheduler.schedulers.blocking import BlockingScheduler
 from sqlmodel import Session, select
 
-from app.ai import execute_agent
-from app.config import config
-from app.db import Agent, AgentQuota, get_db, init_db
+from app.core.ai import execute_agent
+from app.config.config import config
+from app.models.db import Agent, AgentQuota, get_db, init_db
 
 logger = logging.getLogger(__name__)
