+++ docker-compose.yml
@@ -31,7 +31,7 @@ services:
       - OPENAI_API_KEY=${OPENAI_API_KEY}
     ports:
       - "8000:8000"
-    command: uvicorn app.main:app --host 0.0.0.0 --port 8000
+    command: uvicorn app.entrypoints.api:app --host 0.0.0.0 --port 8000
 
   autonomous:
     image: crestal/intentkit:latest
@@ -47,7 +47,7 @@ services:
       - DB_PORT=5432
       - DB_NAME=${POSTGRES_DB:-intentkit}
       - OPENAI_API_KEY=${OPENAI_API_KEY}
-    command: python -m app.autonomous
+    command: python -m app.entrypoints.autonomous
 
 volumes:
   postgres_data:
