+++ requirements.txt
@@ -18,6 +18,7 @@ langchain>=0.3.13
 langchain-core>=0.3.28
 langchain-openai>=0.2.11
 langchain-postgres>=0.0.12
+langchain-community>=0.0.10
 langgraph>=0.2.56
 langgraph-checkpoint>=2.0.8
 langgraph-checkpoint-postgres>=2.0.8
