+++ skills/crestal/__init__.py
@@ -1,6 +1,6 @@
 from langchain_core.tools import BaseTool
 
-from skill.crestal.search_web3_services import search_web3_services
+from skills.crestal.search_web3_services import search_web3_services
 
 
 def get_crestal_skill(name: str) -> BaseTool:
